Separate AGP tests from secondary integration tests
