+++ .github/workflows/main.yml
@@ -16,9 +16,9 @@ jobs:
       fail-fast: false
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        test-suite: [ api-check, unit-test, integration-test-primary, integration-test-secondary, gradle-check ]
+        test-suite: [ api-check, unit-test, integration-test-primary, integration-test-secondary, integration-test-agp, gradle-check ]
     runs-on: ${{ matrix.os }}
-    continue-on-error: ${{ matrix.test-suite == 'integration-test-secondary' }}
+    continue-on-error: ${{ matrix.test-suite == 'integration-test-secondary' || matrix.test-suite == 'integration-test-agp' }}
     defaults:
       run:
         shell: bash
@@ -86,9 +86,13 @@ jobs:
       if: matrix.test-suite == 'integration-test-secondary'
       run: ./gradlew :integration-tests:secondaryTest --stacktrace --info --max-workers=1 --no-parallel -Dorg.gradle.vfs.watch=false -Dorg.gradle.jvmargs="-Xmx20248m -Dkotlin.daemon.jvm.options=-Xmx2048m"
 
+    - name: Run agp integration tests
+      if: matrix.test-suite == 'integration-test-agp'
+      run: ./gradlew :integration-tests:agpTest --stacktrace --info --max-workers=1 --no-parallel -Dorg.gradle.vfs.watch=false -Dorg.gradle.jvmargs="-Xmx20248m -Dkotlin.daemon.jvm.options=-Xmx2048m"
+
     - name: Run remaining tests and checks
       if: matrix.test-suite == 'gradle-check'
-      run: ./gradlew check -x :kotlin-analysis-api:test -x :integration-tests:primaryTest -x :integration-tests:secondaryTest --stacktrace --info --max-workers=1 --no-parallel -Dorg.gradle.vfs.watch=false -Dorg.gradle.jvmargs="-Xmx20248m -Dkotlin.daemon.jvm.options=-Xmx2048m"
+      run: ./gradlew check -x :kotlin-analysis-api:test -x :integration-tests:primaryTest -x :integration-tests:secondaryTest -x :integration-tests:agpTest --stacktrace --info --max-workers=1 --no-parallel -Dorg.gradle.vfs.watch=false -Dorg.gradle.jvmargs="-Xmx20248m -Dkotlin.daemon.jvm.options=-Xmx2048m"
 
     - name: Upload test results
       if: always()
